prod/terraform/modules/get-race_prediction/src_gcf-race_prediction/requirements.txt
pandas==2.2.2
pyppeteer==2.0.0
google-cloud-bigquery==3.14.1
google-cloud-storage==2.17.0
google-cloud-scheduler==2.13.4
google-cloud-secret-manager==2.20.1
grpcio==1.64.1
scikit-learn==1.5.1
lightgbm==4.4.0
pyarrow==17.0.0
google-api-python-client==2.141.0
slack-sdk==3.31.0
